refactor: nonEmptyString predicate and un-barrel
